Remove runner jar
