+++ src/scala/org/pantsbuild/zinc/compiler/BUILD
@@ -24,9 +24,3 @@ scala_library(
   strict_deps=True,
   platform='java8',
 )
-
-jvm_binary(
-  name = "runner",
-  dependencies = [":compiler"],
-  main = "org.pantsbuild.zinc.compiler.Main",
-)
